Added axios.min.js and imported correctly api-client file.
