+++ public/index.html
@@ -1,31 +1,32 @@
 <!DOCTYPE html>
-<html>
+<html lang='en'>
 
 <head>
     <title>HandShake</title>
-    <meta name="author" content="Sofia d'Atri">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="preconnect" href="https://fonts.googleapis.com">
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-    <link href="https://fonts.googleapis.com/css2?family=Francois+One&display=swap" rel="stylesheet">
-    <link rel="icon" href="icons/handshake.png" type="image/png" />
+    <meta name='author' content="Sofia d'Atri">
+    <meta name='viewport' content='width=device-width, initial-scale=1.0' />
+    <link rel='preconnect' href='https://fonts.googleapis.com'>
+    <link rel='preconnect' href='https://fonts.gstatic.com' crossorigin>
+    <link href='https://fonts.googleapis.com/css2?family=Francois+One&display=swap' rel='stylesheet'>
+    <link href='icons/handshake.png' rel='icon' type='image/png' />
     <!--<a href="https://www.flaticon.com/free-icons/partnership-handshake" title="Partnership handshake icons">Partnership handshake icons created by DinosoftLabs - Flaticon</a>-->
     <!-- CSS only -->
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/css/bootstrap.min.css" rel="stylesheet"
-        integrity="sha384-rbsA2VBKQhggwzxH7pPCaAqO46MgnOM80zW1RWuH61DGLwZJEdK2Kadq2F9CUG65" crossorigin="anonymous">
+    <link crossorigin='anonymous' href='https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/css/bootstrap.min.css'
+          integrity='sha384-rbsA2VBKQhggwzxH7pPCaAqO46MgnOM80zW1RWuH61DGLwZJEdK2Kadq2F9CUG65' rel='stylesheet'>
     <!-- JavaScript Bundle with Popper -->
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/js/bootstrap.bundle.min.js"
-        integrity="sha384-kenU1KFdBIe4zVF0s0G1M5b4hcpxyD9F7jL+jjXkk+Q2h455rYXK/7HAuoJl+0I4" crossorigin="anonymous"
-        defer></script>
-        
-        <script defer src="./js/api/api-client.js" ></script>
-        <script defer src="./js/api/ejs.min.js" ></script>
-        <script defer src="./js/api/axios.js" ></script>
-        <script defer src="./js/api/routes.js" ></script>
-        <script defer src="./js/views.js" ></script>
+    <script crossorigin='anonymous'
+            defer integrity='sha384-kenU1KFdBIe4zVF0s0G1M5b4hcpxyD9F7jL+jjXkk+Q2h455rYXK/7HAuoJl+0I4'
+            src='https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/js/bootstrap.bundle.min.js'></script>
 
-    <link href="css/variables.css" rel="stylesheet" type="text/css" />
-    <link href="css/style.css" rel="stylesheet" type="text/css" />
+    <script defer src='./js/api/axios.min.js'></script>
+    <script defer src='./js/api/api-client.js'></script>
+
+    <script defer src='./js/api/ejs.min.js'></script>
+    <script defer src='./js/api/routes.js'></script>
+    <script defer src='./js/views.js'></script>
+
+    <link href='css/variables.css' rel='stylesheet' type='text/css' />
+    <link href='css/style.css' rel='stylesheet' type='text/css' />
 </head>
 
 <body onload = "init()">
